Android performance and bug fixes.
Considerably fewer wormholes are spawned for performance reasons, as they do not perform well on low-end platforms. The Research button should no longer overlap with other buttons on the UI. Fixed the warp shield ability (scout).

+++ src/main/java/org/destinationsol/warp/abilities/WarpShieldAbility.java
@@ -139,8 +139,8 @@ public class WarpShieldAbility implements ShipAbility {
         }
 
         public static AbilityConfig load(JSONObject abNode, ItemManager itemManager, AbilityCommonConfig cc) {
-            float rechargeTime = abNode.getFloat("rechargeTime");
-            float shieldDuration = abNode.getFloat("shieldDuration");
+            float rechargeTime = (float) abNode.getDouble("rechargeTime");
+            float shieldDuration = (float) abNode.getDouble("shieldDuration");
             String warpShieldName = abNode.getString("shield");
             String warpShieldTexName = abNode.optString("shieldTex", warpShieldName);
             itemManager.parseItems(warpShieldName);
